Move user settings to app.py/helpers.py
